+++ src/pel.cpp
@@ -106,13 +106,13 @@ int main(int argc, char* argv[]) {
 
         if (vm.count("evalModel")) {
             LOG(LOG_INFO) << "evaluating model...";
-            unsigned long sum = 0;
+            double sum = 0;
             // evaluate the weight of each formula in the domain
             for(Domain::formula_const_iterator it = d.formulas_begin(); it != d.formulas_end(); it++) {
                 ELSentence formula = *it;
                 //SISet satisfied = d->satisfied(*(formula.sentence()), model);
                 SISet satisfied = formula.sentence()->dSatisfied(model, d);
-                unsigned long weight = d.score(formula, model);
+                double weight = d.score(formula, model);
                 sum += weight;
                 LOG_PRINT(LOG_INFO) << "formula: (" << formula.sentence()->toString() << ")";
                 LOG_PRINT(LOG_INFO) << "\tsatisfied @ " << satisfied.toString();
